Organização de context e drawer na rota de creditors_debtors.
