--- src/components/common/Drawer/creditorDebtor/index.tsx
@@ -1,9 +1,0 @@
-// Imports React
-import React, { Fragment } from 'react'
-
-// Components Imports
-import { GetFormFieldsCreditorDebtor } from './formFields'
-
-export const CreditorDebtorBody: React.FC = () => {
-  return <Fragment>{GetFormFieldsCreditorDebtor()}</Fragment>
-}
